fix: 로또 수익률 계산 오류 수정

+++ src/main/java/lotto/domain/Game.java
@@ -118,7 +118,7 @@ public class Game {
 
     private static void calculateLottoReturns() {
         int totalPrizeMoney = calculateTotalPrizeMoney();
-        lottoReturns = totalPrizeMoney / player.getPurchaseAmount() * 100.0f;
+        lottoReturns = (float) totalPrizeMoney / player.getPurchaseAmount() * 100.0f;
     }
 
     private static Integer calculateTotalPrizeMoney() {
